Move search to MainScene
