+++ app/shared/app-platform/src/commonMain/kotlin/navigation/AniNavigator.kt
@@ -82,7 +82,7 @@ interface AniNavigator {
     }
 
     fun navigateSearch(requestFocus: Boolean = false) {
-        navigator.navigate("/home?tab=search")
+        navigator.navigate("/search?requestFocus=$requestFocus")
     }
 
     /**
